reference_implementations/gcp/step3/vertex/requirements.txt
google-cloud-aiplatform[prediction]==1.55.0
torch==2.2.0
transformers==4.38.1
accelerate==0.27.0
